bug fixed

+++ 金标尺/Controller/Private/LSPrivateChargeViewController.h
@@ -25,7 +25,7 @@ typedef enum {
 @property (nonatomic, strong)NSDate *expireDate;
 @property (nonatomic, strong)NSString *lastDate;
 
-@property (nonatomic, assign)NSInteger totalNum;
+@property (nonatomic, assign)CGFloat totalNum;
 @property (nonatomic, assign)NSInteger price;
 @property (nonatomic, assign)NSInteger kDateSelected;
 @property (nonatomic, assign)NSInteger kPaymentSelected;
